Remove GLAD & GLM from vcpkg
Chkdraft was already using generated glad source files, so either those had to go or the vcpkg dependency had to go; vcpkg is using a three-year-old version of glad which seemed to have no sane way to add singular extensions, so keeping the generated sources. GLM is presently unused and should not be included.