Hacky Remapping
Remapping poses a substantial challenge in OpenGL as there is no great way to get the prior palette index/prior fragment color when rendering a particular image. This hacky approach uses a sum of the colors of the effect as a transparency value which works surprisingly well ableit with unfortunate effects for certain images (dull swarm & psi-field for instance). A better approach should be explored in the future (ping-pong fragment shaders, SSBOs, NV_Texture_Barriers, partially CPU-side rendering, combination of approaches depending on detection of supported OpenGL versions/features).